split part of __init__ to a wrapper.py
