+++ packages/sample-app/sample_app/redis_app.py
@@ -23,8 +23,10 @@ from sentence_transformers import SentenceTransformer
 
 from traceloop.sdk import Traceloop
 from traceloop.sdk.decorators import workflow
+from opentelemetry.sdk.trace.export import ConsoleSpanExporter
 
-Traceloop.init(app_name="redis_app", disable_batch=True)
+Traceloop.init(exporter=ConsoleSpanExporter())
+#Traceloop.init(app_name="redis_app", disable_batch=True)
 
 url = "https://raw.githubusercontent.com/bsbodden/redis_vss_getting_started/main/data/bikes.json"
 response = requests.get(url)
